fix: Cirular dependency on Actor
