+++ src/engine/EmitterType.ts
@@ -1,0 +1,13 @@
+/**
+ * An enum that represents the types of emitter nozzles
+ */
+export enum EmitterType {
+    /**
+     * Constant for the circular emitter type
+     */
+    Circle,
+    /**
+     * Constant for the rectangular emitter type
+     */
+    Rectangle
+}
