+++ src/engine/index.ts
@@ -23,7 +23,9 @@ export * from './Events/MediaEvents';
 export * from './Events';
 export * from './Label';
 export { FontStyle, FontUnit, TextAlign, BaseAlign } from './Graphics/FontCommon';
-export { Particle, ParticleTransform, ParticleEmitter, ParticleArgs, ParticleEmitterArgs, EmitterType } from './Particles';
+export * from './EmitterType';
+export { Particle, ParticleTransform, ParticleArgs, ParticleEmitterArgs } from './Particles';
+export * from './ParticleEmitter';
 export * from './Collision/Physics';
 export * from './Scene';
